cpp06 completo

+++ cpp06/ex01/Serializer.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Serializer.cpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rcastano <rcastano@student.42.fr>          +#+  +:+       +#+        */
+/*   By: roberto <roberto@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 14:10:02 by rcastano          #+#    #+#             */
-/*   Updated: 2024/05/02 18:07:44 by rcastano         ###   ########.fr       */
+/*   Updated: 2024/11/26 14:05:34 by roberto          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 Serializer::Serializer()
 {
-	
+
 }
 
 Serializer::~Serializer()
 {
-	
+
 }
 
 Serializer::Serializer(Serializer &src)
@@ -33,11 +33,11 @@ Serializer& Serializer::operator=(const Serializer &src)
 	return (*this);
 }
 
-int	Serializer::serialize(Data* ptr)
+uintptr_t	Serializer::serialize(Data* ptr)
 {
 	return (reinterpret_cast<uintptr_t>(ptr));
 }
-Data*		Serializer::deserialize(int raw)
+Data*		Serializer::deserialize(uintptr_t raw)
 {
 	return (reinterpret_cast<Data*>(raw));
 }
